pyproject.toml
[tool.poetry]
name = "api-tools"
packages = [
  { include = "gematik" },
]
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.scripts]
gemapi = "gematik.apitools:cli"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
diagrams = "^0.21.1"
pydantic = "^1.9.2"
click = "^8.1.3"
requests = "^2.28.1"
rich = "^12.5.1"
plantuml = "^0.3.0"
igraph = "^0.9.11"
cairocffi = "^1.3.0"
matplotlib = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
